// types/type.d.ts
declare interface ButtonProps extends TouchableOpacityProps {
  title: string;
  className?: string;
  IconRight?: ComponentType<any>;
  IconLeft?: ComponentType<any>;
  bgVariant?: "primary" | "secondary" | "danger" | "success" | "outline";
  textVariant?: "default" | "primary" | "secondary" | "danger" | "success";
}